session: check if event listener is registered before calling Close
The recent addition of session.Close and changes to the event listener
introduced a bug. If no event listener is registered when session.Close
is called, it will panic.

This adds a check on the event listener before trying to close it, and
adds a test case for the bug.

Fixes: https://github.com/strongswan/govici/issues/6

Signed-off-by: Nick Rosbrook <rosbrookn@ainfosec.com>