README: remove old examples and move into getting_started.md
The existing examples do not give much context for how to use the vici
protocol. Remove the old examples and add getting_started.md in an
effort to improve the documentation. This provides examples in a
walkthrough format, and attempts to explain the vici protocol more.

Related: https://github.com/strongswan/govici/issues/8

Signed-off-by: Nick Rosbrook <nr@enr0n.net>