events: fix broken synchronization between listen() and destroy()
Currently, a call to Close or a cancel'd context supplied to Listen does
not synchronize with destroy(). In the case of user cancelling a
context, the event listener's internal context is never cancelled, and
destroyEventListenerWhenClosed never returns. Even if it did, there
would be a race condition with the next call to Listen.

The solution is to add a Context and CancelFunc for use by destroy(),
and a new function called isActive that determines if the event listener
is active by checking if either the destroy or user-supplied context has
been cancel'd. Use this as the check in maybeCreateEventListener, and
remove destroyEventListenerWhenClosed all together.

While working in this area, make sure Session.Close is locked properly
before closing transports.

Add three new integration tests to verify this behavior from now on:
TestListenCancelListenAgain, TestListenWhenAlreadyActive, and
TestCloseWithActiveNextEvent.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>