events: export Event type and expose event name
Currently, the NextEvent API only returns the message associated with
the event, but does not expose the event name, e.g. "ike-updown." This
means that a caller who has registered the event listener for multiple
events has no way to distinguish messages by event type.

To solve this, expose the event name as set by the server in the event
packet. This is done by modifying the old event type to look like:

  type Event struct {
      Name    string
      Message *Message
      err     error
  }

and modifying NextEvent to return this new Event type, rather than
*Message.

Add an integration test to verify the new behvaior.

Closes: https://github.com/strongswan/govici/issues/18

Signed-off-by: Nick Rosbrook <nr@enr0n.net>