events: convert Listen API to Subscribe API
Besides the name change of the API, the new Subscribe API should allow
users to add to their list of subscribed events at any time. This was
not possible before, as subsequent calls to Listen would return an
error. Now, if the event listener is already active, just call
el.registerEvents with the new events.

If an event is already registered, don't re-register it, and don't
return an error. Keep track of events registered to prepare for the
addition of the Unsubscribe and UnsubscribeAll APIs.

Finally, modify the old integration test that ensured subsequent Listen
calls failed to check that subsequent Subscribe calls do NOT fail. Other
tests are left as is, besides their name being changed.

Add an integration test that makes sure that consecutive calls to
Subsribe register only new events, i.e. el.events is the intersection of
all calls to subscribe.

Related: https://github.com/strongswan/govici/issues/19

Signed-off-by: Nick Rosbrook <nr@enr0n.net>