events: modify event listener lifecycle
In order to accomodate the new Subscribe API, re-work the event
listener lifecycle. Start the listen goroutine when the event
listener is initialized in newEventListener. This goroutine is only
stopped when the event listener is Close()'d.

In order to keep packet transmission with the daemon consistent, make it
the responsibility of listen() to handle all event-related packets. If
it gets a pktConfirm or pktUnknown, send it over the new packet channel
for a waiting call to eventTransportCommunicate.

This change allows for the isActive and destroy logic to be removed.

Related: https://github.com/strongswan/govici/issues/19

Signed-off-by: Nick Rosbrook <nr@enr0n.net>