events: tune error handling for new event lifecycle
In the new event subscription model, listen() has additional
responsibilities to report event registration/unregistration responses.
When listen() encounters an error on recv(), it should only notify the
event channel if there are registered events, and should not block if
there is not an active call to eventTransportCommunicate.

Correct this with the perr channel, and checking the length of el.events
before sending the error to the respective callers. The former requires
adding a mutex to protect read/write of el.events, since listen() is in
a separate goroutine from callers to Subscribe/Unsubscribe.

Add a test to make sure the event channel is not notified of errors when
no events are currently registered.

Finally, add error return values to Unsubscribe and UnsubscribeAll.

Related: https://github.com/strongswan/govici/issues/19

Signed-off-by: Nick Rosbrook <nr@enr0n.net>