events: correctly handle io.EOF in listen()
When an event is registered, and the event listener's transport closes
for some reason (e.g. charon is stopped) Session.Close() will never
return because the listen() loop will fill up the event channel with EOF
errors, and block indefinitely on a channel write (therefore never
checking if the event listener context is Done() again).

When the event listener receives io.EOF, it should return. Doing this
fixes the potential deadlock issue and allows Session.Close() to return
in such a case.

Add a regression test to verify this fix.

Fixes: https://github.com/strongswan/govici/issues/24

Signed-off-by: Nick Rosbrook <nr@enr0n.net>