message: define and verify "empty" message elements
Add documentation explaining the definition of "empty message element",
thus setting explicit requirements on emptyMessageElement. Also, add
cases to the switch statement in emptyMessageElement for all supported
integer types, and bool. This has no functional change, but does assist
anyone reading the code.

Add TestEmptyMessageElement that tests emptyMessageElement directly and
serves as a verification of our new "empty message element" definition.
This test reveals that previously a zero-length slice would be
considered as non-empty. Since this is not consistent with the
definition, update emptyMessageElement to return true if a zero-length
slice is given.

Closes: https://github.com/strongswan/govici/issues/30

Signed-off-by: Nick Rosbrook <nr@enr0n.net>