message: do not use chan to implement ordered iterate
The Message.elements() function is used to iterate over the message data
in order, as specified by the message keys. It currently creates a chan
messageElement to return to the caller, and starts another goroutine to
produce on the channel. This works fine, but is a little un-necessary.

Simplify things by just creating a []messageElement, and allowing
callers to iterate over that.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>