events: do not send listen() errors to event channel
Currently, when an unexpected error is encountered in the listen() loop,
if there are registered events, an error is sent over the ec (the chan
Event) for the purpose of notifying NextEvent() callers. However, this
is not useful to callers of NextEvent(), as they only want to know (1)
when a valid event is received, and (2) when no more events can be
received.

For this reason, the current logic adds unnecessary complexity to the
listen() error handling. Simply remove this logic since it does not
provide any real benefit to package users or internally.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>