events: do not call SetReadDeadline() in listen()
In 6dc3797a391a695900ec6188de494028c4fcda48, a read deadline was added
so that listen() would not block forever on recv() calls when the
listener was closed. This was necessary because the event listener was
Close()'d by cancelling a context. If recv() was blocking, the loop
would not ever notice the closed context.

In the current logic, Close() will call el.conn.Close() which is all we
really need to stop the event loop. To do this, however, just return on
any error from recv(). With this, strip out the perr logic as there is
no longer a need for it. And, strip out the ctx and cancel logic because
as noted, closing the connection and returning on error from recv() is
sufficient.

Overall, the reduction of complexity should make this code easier to
maintain going forward.

Related: https://github.com/strongswan/govici/issues/34

Signed-off-by: Nick Rosbrook <nr@enr0n.net>