vici: ensure listen() cannot get stuck on event channel
Currently, the write to the event channel will block until there is room
in the buffer. If a package user calls Subscribe, but does not call
NextEvent enough to keep room in the buffer, the listen() loop can get
stuck.

And, if a caller used the recently added NotifyEvents method, NextEvent
may never get called, which would eventually mean their registered
channel would stop receiving events too.

To fix this, change the behavior of NextEvent: when an event is received
and the buffer is full, discard the oldest event to make room for the
new one. Update the NextEvent documentation to reflect this change.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>