vici: use notify() for internal event channel
Currently, the write to the event channel will block until there is room
in the buffer. If a package user calls Subscribe, but does not call
NextEvent enough to keep room in the buffer, the listen() loop can get
stuck.

And, if a caller used the recently added NotifyEvents method, NextEvent
may never get called, which would eventually mean their registered
channel would stop receiving events too.

To fix this, make the internal event channel one of the registered
channels. However, this does change the behavior of NextEvent: when the
buffer is full, new events will be discarded, just like with the use of
NotifyEvents.

Signed-off-by: Nick Rosbrook <nr@enr0n.net>