all: re-work client connection to be context-aware
The current API does not accept context.Context in functions that may be
blocking on network I/O, and the underlying transport is not written in
a way that this could easily be added.

Re-write the underlying packet connection logic using a new clientConn
type. This new type has similar functionality to the previous transport
type, but is built with context support.

Adapt the session and event listener code to use this new type, but just
pass context.Background() for now. The existing API can be changed to
accept context.Context later, or "Context" variants can be added.