session: implement Call and CallStreaming
Now that the internals are context-aware, expose API on Session that
allows passing a context when making command requests to the server.
In particular, add Call and CallStreaming to the API.

Call is just like CommandRequest, but (a) taking the opportunity for a
better name, and (b) accepts a context. CallStreaming is the same ways
derived from StreamedCommandRequest, but has additional differences.

Namely, CallStreaming makes use of the new iter.Seq iterator pattern.
Instead of blocking until all the reponse messages have been received
from the server, return an iterator once the initial command request
and event registration have succeeded. Then, just yield response
messages from the server as they are received. The provided context
can be used to interrupt any of the network I/O invovled in the command
request and subsequent event streaming.