vici: add unified client conn implementation
Add a new "unified" implementation of a client conn. It is unified in
the sense that one conn is used for command requests and event
listening.

The basic structure is that one goroutine is running the listen() loop,
which is responsible for reading all data from the server. For regular
requests, the response is then written to an internal packet chan.
Events are dispatched to subscribers, and in the case of streaming
events, to an internal event chan.

This structure allows a single context-aware client conn for all request
types. Read and write sequence counters are maintained to match
responses with the correct response in the event of writer abandoning
a response. If the read sequence surpasses the write sequence, it is a
server error. Event packets do not increment the read sequence.

A new test suite is added which makes some use of the experimental
synctest package to simplify the testing.

Later commits will make use of this in the Session type.