client-conn: remove internal event chan
The current streaming command logic has a flaw, in that the select block
in wait() may yield the pktCmdResponse packet from a streaming command
before yielding all of the associated event packets. This is intentional
in the design of select, and we cannot rely on the ordering to be
deterministic.

To fix this, simply remove the internal event chan, and send the event
message over the usual packet chan. But, do not increment the write
sequence when reading event packets in wait().

Fixes: https://github.com/strongswan/govici/issues/50