test: fuzz packet decode/encode with FuzzEncodeDecode
Decoding raw bytes over a socket is a good candidate for fuzzing, so add
some with a new FuzzEncodeDecode function. For now, the fuzz input is
really the raw data to decode, but we verify that unless an error is
returned, any packet is valid, and re-encoding yields the original data.

(cherry picked from commit af577804d566c5210d81d29ecee7d098e6b5e6bb)