message: refuse to overwrite existing key in decode()
This is undefined behavior, and should be refused during decoding. Add
an addItemUnique helper that returns an error when addItem would
overwrite.

This case might need to be addressed in marshaling, too. But, it's not
always an error, because e.g. Message.Set should be allowed to overwrite.

Detected by FuzzEncodeDecode/ef9573a1b1a1c540.