all: add sendmsg() and readmsg() helpers
The basic logic for reading and writing packets to a net.Conn is
repeated in tests, and arguably belongs in message.go anyways; it's just
an extra step of reading/writing the packet length from/to the conn.

Add some basic helpers in the message.go code where the rest of this
logic lives. This allows simplifying clientConn.write() and removing
clientConn.read() all together. And, these can be used by the test code
instead of duplicating read()/write().

The TestClientConnRead and TestClientConnWrite tests are now a bit
awkward, but this will be addressed in a later commit.